+++ README.md
@@ -123,9 +123,8 @@
   - [x] (프로덕트 코드) `button` 태그 삭제 -> 테스트 실패
   - [x] (테스트 실패를 성공시키기 위해) js를 이용해 동적으로 `button` 태그 제작 -> 테스트 성공
   - [x] `button` 태그를 컴포넌트로 제작
-  - [ ] `button` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
-    - [ ] 동적 인자 ->`button`에 들어갈 타입 및 들어갈 이름
-    - [ ] `button` 테스트 코드 수정 (취소하기, 등록하기)
+  - [x] `button` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
+    - [x] 동적 인자 ->`button`에 들어갈 타입 및 들어갈 이름
 
 - [ ] `modal-container` (.modal-container) => 흰색 바탕 부분 (backdrop 제외)
